refactor: migrate from @sentry/minimal to @sentry/node for improved error handling
