+++ backend/package.json
@@ -48,7 +48,6 @@
 		"@nestjs/throttler": "^6.5.0",
 		"@nestjs/typeorm": "^11.0.1",
 		"@rocketadmin/shared-code": "workspace:*",
-		"@sentry/minimal": "^6.19.7",
 		"@sentry/node": "10.53.1",
 		"@toon-format/toon": "^2.3.0",
 		"@types/crypto-js": "^4.2.2",
